websockets/opening-handshake/002.html

<!doctype html>
<title>WebSockets: valid handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
setup({timeout:10010});
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/ws/echo');
  ws.onopen = t.step_func(function(e) {    
    ws.onclose = t.step_func(function(e) {
      assert_equals(e.wasClean, true);
      ws.onclose = t.step_func(function() {assert_unreached()});  
      setTimeout(t.step_func(function() {t.done();}), 50)
    })
    ws.close();
  });
  ws.onerror = ws.onmessage = ws.onclose = t.step_func(function() {assert_unreached()});
}, null, {timeout:10000});
</script>
